+++ .eslintignore
@@ -1,2 +1,3 @@
 /.nuxt
 /.output
+/.tsup-output
