+++ .gitignore
@@ -4,3 +4,4 @@ node_modules
 nuxt.d.ts
 .output
 .env
+.tsup-output
