+++ package.json
@@ -1,22 +1,28 @@
 {
+  "dependencies": {
+    "express": "^4.18.2"
+  },
   "devDependencies": {
     "@aws-sdk/client-s3": "^3.281.0",
     "@marp-team/marp-cli": "^2.4.0",
     "@nuxt/eslint-config": "^0.1.1",
     "animate.css": "^4.1.1",
+    "cors": "^2.8.5",
+    "dotenv-cli": "^7.0.0",
     "eslint": "^8.35.0",
     "eslint-config-prettier": "^8.6.0",
-    "express": "^4.18.2",
     "npm-check-updates": "^16.7.10",
     "npm-run-all": "^4.1.5",
     "nuxt": "^3.2.3",
-    "prettier": "^2.8.4"
+    "prettier": "^2.8.4",
+    "tsup": "^6.6.3"
   },
   "packageManager": "npm@8.11.0",
   "private": true,
   "scripts": {
     "build": "nuxi build",
     "dev": "nuxi dev",
+    "dev:express": "tsup express.js --onSuccess=\"dotenv -e .env -- node .tsup-output/express.js\"",
     "format": "prettier --write .",
     "lint": "eslint --fix .",
     "slide": "marp -s docs",
