+++ server/api/[...].js
@@ -1,131 +1,5 @@
-import express from "express";
-import {
-  findTrainers,
-  findTrainer,
-  upsertTrainer,
-  deleteTrainer,
-} from "~/server/utils/trainer";
-import { findPokemon } from "~/server/utils/pokemon";
-
-const app = express();
-
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
-const router = express.Router();
-
-router.get("/hello", (_req, res) => {
-  res.send("Hello World");
-});
-
-/** トレーナー名の一覧の取得 */
-router.get("/trainers", async (_req, res, next) => {
-  try {
-    const trainers = await findTrainers();
-    const trainerNames = trainers.map(({ Key }) => Key.replace(/\.json$/, ""));
-    res.send(trainerNames);
-  } catch (err) {
-    next(err);
-  }
-});
-
-/** トレーナーの追加 */
-router.post("/trainer", async (req, res, next) => {
-  try {
-    if (!("name" in req.body && req.body.name.length > 0))
-      return res.sendStatus(400);
-    const trainers = await findTrainers();
-    if (trainers.some(({ Key }) => Key === `${req.body.name}.json`))
-      return res.sendStatus(409);
-    const result = await upsertTrainer(req.body.name, req.body);
-    res.status(result["$metadata"].httpStatusCode).send(result);
-  } catch (err) {
-    next(err);
-  }
-});
-
-/** トレーナーの取得 */
-router.get("/trainer/:trainerName", async (req, res, next) => {
-  try {
-    const { trainerName } = req.params;
-    const trainer = await findTrainer(trainerName);
-    res.send(trainer);
-  } catch (err) {
-    next(err);
-  }
-});
-
-/** トレーナーの更新 */
-router.post("/trainer/:trainerName", async (req, res, next) => {
-  try {
-    const { trainerName } = req.params;
-    const trainers = await findTrainers();
-    if (!trainers.some(({ Key }) => Key === `${trainerName}.json`))
-      return res.sendStatus(404);
-    const result = await upsertTrainer(trainerName, req.body);
-    res.status(result["$metadata"].httpStatusCode).send(result);
-  } catch (err) {
-    next(err);
-  }
-});
-
-/** トレーナーの削除 */
-router.delete("/trainer/:trainerName", async (req, res, next) => {
-  try {
-    const { trainerName } = req.params;
-    const result = await deleteTrainer(trainerName);
-    res.status(result["$metadata"].httpStatusCode).send(result);
-  } catch (err) {
-    next(err);
-  }
-});
-
-/** ポケモンの追加 */
-router.put(
-  "/trainer/:trainerName/pokemon/:pokemonName",
-  async (req, res, next) => {
-    try {
-      const { trainerName, pokemonName } = req.params;
-      const trainer = await findTrainer(trainerName);
-      const pokemon = await findPokemon(pokemonName);
-      const {
-        order,
-        name,
-        sprites: { front_default },
-      } = pokemon;
-      trainer.pokemons.push({
-        id: (trainer.pokemons[trainer.pokemons.length - 1]?.id ?? 0) + 1,
-        nickname: "",
-        order,
-        name,
-        sprites: { front_default },
-      });
-      const result = await upsertTrainer(trainerName, trainer);
-      res.status(result["$metadata"].httpStatusCode).send(result);
-    } catch (err) {
-      next(err);
-    }
-  }
-);
-
-/** ポケモンの削除 */
-router.delete(
-  "/trainer/:trainerName/pokemon/:pokemonId",
-  async (req, res, next) => {
-    try {
-      const { trainerName, pokemonId } = req.params;
-      const trainer = await findTrainer(trainerName);
-      const index = trainer.pokemons.findIndex(
-        (pokemon) => String(pokemon.id) === pokemonId
-      );
-      trainer.pokemons.splice(index, 1);
-      const result = await upsertTrainer(trainerName, trainer);
-      res.status(result["$metadata"].httpStatusCode).send(result);
-    } catch (err) {
-      next(err);
-    }
-  }
-);
+import app from "~/server/utils/app";
+import router from "~/server/utils/router";
 
 app.use("/api", router);
