+++ server/utils/app.js
@@ -1,0 +1,8 @@
+import express from "express";
+
+const app = express();
+
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+export default app;
