+++ tsup.config.js
@@ -1,0 +1,6 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  inject: ["tsup-shim.js"],
+  outDir: ".tsup-output",
+});
